requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
htmx==0.4.1
gunicorn==21.2.0
weasyprint==60.2

# Database and Cache
sqlalchemy==2.0.27
mysqlclient==2.2.4
alembic==1.13.1
redis==5.0.1
celery==5.3.6

# Authentication and Security
pyseto==1.7.5
bcrypt==4.1.2
authlib==1.3.0
pyotp==2.9.0
python-fido2==1.1.3
hibp==3.1.0
python-jose==3.3.0
slowapi==0.1.9
clamav==1.2.0

# ML/AI
transformers==4.38.1
spacy==3.7.4
torch==2.2.1
vllm==0.3.3
llama-cpp-python==0.2.56
sentence-transformers==2.5.1
faiss-cpu==1.8.0  # or faiss-gpu
onnxruntime==1.17.0
detoxify==0.5.1
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351

# Web Search
searx==1.0.0
wikipedia-api==0.6.0
arxiv==2.1.0

# Vector Store
qdrant-client==1.7.3

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.44b0
prometheus-client==0.19.0
grafana-loki-client==0.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
requests==2.31.0
aiohttp==3.9.3
minio==7.2.5
pygments==2.17.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
locust==2.24.0
playwright==1.42.0
bandit==1.7.7
mypy==1.8.0
ruff==0.2.2
black==24.2.0

# Deployment
docker-compose==1.29.2